fix: registration for mobile

+++ user/views.py
@@ -49,7 +49,8 @@ class RegisterViewSet(viewsets.ViewSet):
                 status_code=400
             )
 
-        user.delete()
+        if user and user.status == "deleted":
+            user.delete()
 
         serializer.validated_data.pop("password2", None)
         user = User.objects.create_user(**serializer.validated_data)
@@ -305,7 +306,7 @@ class SendPasswordResetOtpView(GenericAPIView):
             )
         
         EmailUtil.send_password_reset_email(user)
-
+        
         return CustomResponse.success(
             message="Password reset otp has been sent",
             status_code=200
